Big ugly refactor part 2

+++ clownaudio/clownaudio.c
@@ -8,11 +8,11 @@
 
 static BackendStream *stream;
 
-static void CallbackStream(void *user_data, float *output_buffer, unsigned long frames_to_do)
+static void CallbackStream(void *user_data, float *output_buffer, size_t frames_to_do)
 {
 	(void)user_data;
 
-	for (unsigned long i = 0; i < frames_to_do * STREAM_CHANNEL_COUNT; ++i)
+	for (size_t i = 0; i < frames_to_do * STREAM_CHANNEL_COUNT; ++i)
 		output_buffer[i] = 0.0f;
 
 	Mixer_MixSamples(output_buffer, frames_to_do);
