+++ clownaudio/decode/predecoder.h
@@ -13,4 +13,4 @@ void Predecoder_UnloadData(PredecoderData *data);
 Predecoder* Predecoder_Create(PredecoderData *data, bool loop, DecoderInfo *info);
 void Predecoder_Destroy(Predecoder *predecoder);
 void Predecoder_Rewind(Predecoder *predecoder);
-unsigned long Predecoder_GetSamples(Predecoder *predecoder, void *buffer, unsigned long frames_to_do);
+size_t Predecoder_GetSamples(Predecoder *predecoder, void *buffer, size_t frames_to_do);
