+++ clownaudio/decode/split_decoder.h
@@ -11,5 +11,5 @@ void SplitDecoder_UnloadData(SplitDecoderData *data);
 SplitDecoder* SplitDecoder_Create(SplitDecoderData *data, bool loop, unsigned long sample_rate);
 void SplitDecoder_Destroy(SplitDecoder *split_decoder);
 void SplitDecoder_Rewind(SplitDecoder *split_decoder);
-unsigned long SplitDecoder_GetSamples(SplitDecoder *split_decoder, void *buffer, unsigned long frames_to_do);
+size_t SplitDecoder_GetSamples(SplitDecoder *split_decoder, void *buffer, size_t frames_to_do);
 void SplitDecoder_SetSampleRate(SplitDecoder *split_decoder, unsigned long sample_rate1, unsigned long sample_rate2);
