+++ clownaudio/mixer.h
@@ -21,4 +21,4 @@ void Mixer_CancelFade(Mixer_Sound instance);
 void Mixer_SetSoundVolume(Mixer_Sound instance, float volume);
 void Mixer_SetSoundSampleRate(Mixer_Sound instance, unsigned long sample_rate1, unsigned long sample_rate2);
 void Mixer_SetPan(Mixer_Sound instance, float pan);	// -1.0f is full-left; 1.0f is full-right
-void Mixer_MixSamples(float *output_buffer, unsigned long frames_to_do);
+void Mixer_MixSamples(float *output_buffer, size_t frames_to_do);
