+++ clownaudio/playback/miniaudio.c
@@ -8,7 +8,7 @@
 
 struct BackendStream
 {
-	void (*user_callback)(void*, float*, unsigned long);
+	void (*user_callback)(void*, float*, size_t);
 	void *user_data;
 
 	ma_device device;
@@ -40,7 +40,7 @@ void Backend_Deinit(void)
 	
 }
 
-BackendStream* Backend_CreateStream(void (*user_callback)(void*, float*, unsigned long), void *user_data)
+BackendStream* Backend_CreateStream(void (*user_callback)(void*, float*, size_t), void *user_data)
 {
 	BackendStream *stream = malloc(sizeof(BackendStream));
