+++ clownaudio/playback/playback.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <stdbool.h>
+#include <stddef.h>
 
 #define STREAM_SAMPLE_RATE 48000
 #define STREAM_CHANNEL_COUNT 2
@@ -9,7 +10,7 @@ typedef struct BackendStream BackendStream;
 
 bool Backend_Init(void);
 void Backend_Deinit(void);
-BackendStream* Backend_CreateStream(void (*callback)(void*, float*, unsigned long), void *user_data);
+BackendStream* Backend_CreateStream(void (*callback)(void*, float*, size_t), void *user_data);
 bool Backend_DestroyStream(BackendStream *stream);
 bool Backend_SetVolume(BackendStream *stream, float volume);
 bool Backend_PauseStream(BackendStream *stream);
