+++ clownaudio/playback/portaudio.c
@@ -8,7 +8,7 @@
 
 struct BackendStream
 {
-	void (*user_callback)(void*, float*, unsigned long);
+	void (*user_callback)(void*, float*, size_t);
 	void *user_data;
 
 	PaStream *pa_stream;
@@ -44,7 +44,7 @@ void Backend_Deinit(void)
 	Pa_Terminate();
 }
 
-BackendStream* Backend_CreateStream(void (*user_callback)(void*, float*, unsigned long), void *user_data)
+BackendStream* Backend_CreateStream(void (*user_callback)(void*, float*, size_t), void *user_data)
 {
 	BackendStream *stream = malloc(sizeof(BackendStream));
